Begin refactor of pkg/server, move to pkg/old_server
